Немного поправил код

+++ main.py
@@ -1,8 +1,8 @@
 import os
 import sys
-
+from typing import List
 import requests
-from PyQt5 import uic, QtCore
+from PyQt5 import uic, QtCore, QtGui
 from PyQt5.QtGui import QPixmap
 from PyQt5.QtWidgets import QApplication, QMainWindow, QLineEdit
 
@@ -16,35 +16,26 @@ class Main(QMainWindow):
         self.position = [37.615370, 55.756936]
         self.scale = 5
 
-    def get_map(self):
-        url_template = 'http://static-maps.yandex.ru/1.x/?ll= &spn= &l=map'
-        self.lineEdit.setText(self.lineEdit.text().strip())
-        try:
-            os.remove(self.map_file)
-            self.map_file = None
-        except TypeError:
-            pass
+    def isEmpty(self, lineEdit: QLineEdit):
+        if lineEdit.text() == '':
+            lineEdit.setStyleSheet('border: 1px solid red;')
+            return True
+        else:
+            lineEdit.setStyleSheet('border: 1px solid rgb(160, 160, 160);')
+        return False
 
-        def isEmpty(lineEdit: QLineEdit):
-            """Проверка на то, что строка не пустая."""
-            if lineEdit.text() == '':
-                lineEdit.setStyleSheet('border: 1px solid red;')
-                return True
-            else:
-                lineEdit.setStyleSheet('border: 1px solid rgb(160, 160, 160);')
-            return False
-
-        if not (isEmpty(self.lineEdit) and isEmpty(self.lineEdit_2)):
-            coord = [','.join(self.lineEdit.text().split(' '))]
-            self.position = list(map(float, self.lineEdit.text().split(' ')))
-            if len(self.lineEdit_2.text().strip().split(' ')) == 1:
-                coord.append(','.join([self.lineEdit_2.text().strip(), self.lineEdit_2.text().strip()]))
-                self.scale = float(self.lineEdit_2.text().strip())
-            else:
-                max_scale = str(max(list(map(float, self.lineEdit_2.text().strip().split(' ')))))
-                self.lineEdit_2.setText(max_scale)
-                self.scale = float(max_scale)
-                coord.append(','.join([max_scale, max_scale]))
+    def get_map(self, pos: List, scale: str):
+        if (pos and scale) is not None:
+            url_template = 'http://static-maps.yandex.ru/1.x/?ll= &spn= &l=map'
+            try:
+                os.remove(self.map_file)
+                self.map_file = None
+            except TypeError:
+                pass
+            coord = [','.join(pos)]
+            self.position = list(map(float, pos))
+            coord.append(','.join([scale, scale]))
+            self.scale = int(scale)
             for i in coord:
                 url_template = url_template.replace(' ', i, 1)
             response = requests.get(url_template)
@@ -57,60 +48,70 @@ class Main(QMainWindow):
             pixmap = QPixmap(self.map_file)
             self.label.setPixmap(pixmap)
 
+    def data_prepare(self, lineEdit: QLineEdit, key=None):
+        """Возвращает список из данных, которые были в lineEdit.
+        Если передан ключ, то передано значение scale."""
+        if not self.isEmpty(lineEdit):
+            if key is None:
+                if len(lineEdit.text().split(',')) == 1:
+                    data = lineEdit.text().split()
+                else:
+                    data = list(map(lambda x: x.strip(), lineEdit.text().split(',')))
+                lineEdit.setText(lineEdit.text().strip())
+            else:
+                if lineEdit.text().split(',') == 1:
+                    lineEdit.setText(str(max(list(map(int, lineEdit.text().split())))))
+                else:
+                    lineEdit.setText(str(max(list(map(int,
+                                                      list(map(lambda x: x.strip(),
+                                                               lineEdit.text().split(','))))))))
+                data = lineEdit.text()
+        else:
+            data = None
+        return data
+
     def closeEvent(self, event):
         try:
             os.remove(self.map_file)
         except TypeError:
             pass
 
-    def mousePressEvent(self, event):
-        focused_widget = QApplication.focusWidget()
-        if isinstance(focused_widget, QLineEdit):
-            focused_widget.clearFocus()
-        QMainWindow.mousePressEvent(self, event)
-
     def keyPressEvent(self, key):
-        """Обработка нажатия различных клавиш."""
         focused_widget = QApplication.focusWidget()
         if isinstance(focused_widget, QLineEdit):
             focused_widget.clearFocus()
         if key.key() in [QtCore.Qt.Key_Enter, QtCore.Qt.Key_Return]:
-            self.get_map()
+            self.get_map(self.data_prepare(self.lineEdit), self.data_prepare(self.lineEdit_2, key=True))
         if key.key() == QtCore.Qt.Key_PageUp:
             if self.scale < 90:
                 self.scale += 1
             self.lineEdit_2.setText(str(round(self.scale)))
-            self.get_map()
+            self.get_map(self.data_prepare(self.lineEdit), self.data_prepare(self.lineEdit_2, key=True))
         if key.key() == QtCore.Qt.Key_PageDown:
             if self.scale > 0:
                 self.scale -= 1
-            print(self.scale)
             self.lineEdit_2.setText(str(round(self.scale)))
-            self.get_map()
+            self.get_map(self.data_prepare(self.lineEdit), self.data_prepare(self.lineEdit_2, key=True))
         if key.key() == QtCore.Qt.Key_Up:
-            if int(self.position[1]) < 70:
+            if float(self.position[1]) < 70:
                 self.position[1] += 0.5
-            print(self.position)
             self.lineEdit.setText(' '.join(list(map(str, self.position))))
-            self.get_map()
+            self.get_map(self.data_prepare(self.lineEdit), self.data_prepare(self.lineEdit_2, key=True))
         if key.key() == QtCore.Qt.Key_Down:
-            if int(self.position[1]) > -70:
+            if float(self.position[1]) > -70:
                 self.position[1] -= 0.5
-            print(self.position)
             self.lineEdit.setText(' '.join(list(map(str, self.position))))
-            self.get_map()
+            self.get_map(self.data_prepare(self.lineEdit), self.data_prepare(self.lineEdit_2, key=True))
         if key.key() == QtCore.Qt.Key_Right:
             if int(self.position[0]) < 180:
                 self.position[0] += 0.5
-            print(self.position)
             self.lineEdit.setText(' '.join(list(map(str, self.position))))
-            self.get_map()
+            self.get_map(self.data_prepare(self.lineEdit), self.data_prepare(self.lineEdit_2, key=True))
         if key.key() == QtCore.Qt.Key_Left:
             if int(self.position[0]) > 0:
                 self.position[0] -= 0.5
-            print(self.position)
             self.lineEdit.setText(' '.join(list(map(str, self.position))))
-            self.get_map()
+            self.get_map(self.data_prepare(self.lineEdit), self.data_prepare(self.lineEdit_2, key=True))
 
 
 if __name__ == '__main__':
